Move ArchiveEntry, ArchiveEntryDataReader and ArchiveEntryId to an
entry module
