+++ bindings/C/src/lib.rs
@@ -10,10 +10,10 @@ use mla::crypto::mlakey::{
     parse_mlakey_privkey_der, parse_mlakey_privkeys_pem_many, parse_mlakey_pubkey_der,
     parse_mlakey_pubkeys_pem_many,
 };
+use mla::entry::ArchiveEntryId;
 use mla::errors::ConfigError;
 use mla::errors::Error as MLAError;
 use mla::helpers::linear_extract;
-use mla::ArchiveEntryId;
 use mla::ArchiveReader;
 use mla::ArchiveWriter;
 use std::collections::HashMap;
